requirements.txt
requests==2.23.0
APScheduler==3.6.3
Django==3.0.6
djangorestframework==3.11.0
beautifulsoup4==4.9.1
lxml==4.5.0
